Update
Tried for ages to figure out getting a rectangle to fit the sheet. Depends upon the width of the trace but behaves funny at 63pixels wide. Made the canvas larger than required so as not to have a problem. Fixed bug with centering the text - added fudge factor but not sure why it needs it.